Refactored model into the api module, it no longer makes sense to have it as a separate module
